requirements.txt
opencv-contrib-python==4.9.0.80
numpy==1.26.4 
opencv-python == 4.9.0.80
pillow==10.2.0
streamlit==1.32.0
streamlit-webrtc==0.47.6
streamlit-card-1.0.0